// src/types.ts
export enum operations {
  mouse_up = 'mouse_up',
  mouse_down = 'mouse_down',
  mouse_left = 'mouse_left',
  mouse_right = 'mouse_right',
  mouse_position = 'mouse_position',
  draw_circle = 'draw_circle',
  draw_rectangle = 'draw_rectangle',
  draw_square = 'draw_square',
  prnt_scrn = 'prnt_scrn'
}